[REFACTOR] change file name notation
