+++ SSD/file_handler_test.cpp
@@ -1,5 +1,6 @@
 #include "gmock/gmock.h"
-#include "FileHandler.h"
+#include "file_handler.h"
+
 TEST(FileHandler, DISABLED_rename) {
 	FileHandler fileHandler;
 	EXPECT_NO_THROW(fileHandler.rename("aaa.txt", "bbb.txt"));
