+++ SSD/nand_reader.h
@@ -3,7 +3,7 @@
 #include <string>
 #include "nand_flash_memory.h"
 #include "buffer_manager.h"
-#include "SSDCommand.h"
+#include "ssd_command.h"
 
 using std::string;
 class NandReader : public SSDCommand {
