+++ SSD/nand_writer.h
@@ -3,7 +3,7 @@
 #include "nand_flash_memory.h"
 #include "buffer_manager.h"
 #include "global_config.h"
-#include "SSDCommand.h"
+#include "ssd_command.h"
 using std::string;
 
 class NandWriter : public SSDCommand
